% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svfgrid.R
\name{calculate_df_grid}
\alias{calculate_df_grid}
\title{Calcular el dataframe de grid para SVFGrid}
\usage{
calculate_df_grid(grid)
}
\arguments{
\item{grid}{Objeto SVFGrid sobre el cual operar.}
}
\value{
El objeto SVFGrid con el dataframe de grid actualizado.
}
\description{
Este método calcula y añade información adicional al dataframe de grid
asociado a un objeto SVFGrid. Esta información incluye los valores de phi
para cada celda del grid y las celdas contiguas a cada celda.
}
