% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svfgrid.R
\name{calculate_dmu_phi}
\alias{calculate_dmu_phi}
\title{Calcular phi para una celda específica en SVFGrid}
\usage{
calculate_dmu_phi(grid, cell)
}
\arguments{
\item{grid}{Objeto SVFGrid sobre el cual operar.}

\item{cell}{Vector que especifica la posición de la celda en el grid.}
}
\value{
Una lista que contiene los valores de phi para la celda especificada.
}
\description{
Esta función calcula y retorna el valor de phi para una celda específica del
grid, basado en los datos del grid y la posición de la celda.
}
