% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svfgrid.R
\name{create_grid.SVFGrid}
\alias{create_grid.SVFGrid}
\title{Crear grid para SVFGrid}
\usage{
create_grid.SVFGrid(grid)
}
\arguments{
\item{grid}{Objeto SVFGrid sobre el cual operar.}
}
\value{
El objeto SVFGrid con el grid creado.
}
\description{
Este método crea un grid basado en los datos y el parámetro d proporcionados
al constructor. Este grid es una representación de los datos en un espacio
dividido en celdas definidas por el parámetro d.
}
