% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos2.R
\docType{data}
\name{datos2}
\alias{datos2}
\title{datos2.txt}
\format{
\subsection{\code{datos2}}{

Un data frame con 4 observaciones de 4 variables.

\describe{
\item{\code{x1}}{Variable de entrada 1.}
\item{\code{x2}}{Variable de entrada 2.}
\item{\code{y1}}{Variable de salida 1.}
\item{\code{y2}}{Variable de salida 2.}
}
}
}
\source{
Los datos fueron creados con propósitos ilustrativos.
}
\usage{
datos2
}
\description{
Este conjunto de datos contiene cuatro variables: \code{x1}, \code{x2}, \code{y1} y \code{y2}.
}
\keyword{datasets}
