% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos3.R
\docType{data}
\name{datos3}
\alias{datos3}
\title{datos3.txt}
\format{
\subsection{\code{datos3}}{

Un data frame con 4 observaciones de 3 variables.

\describe{
\item{\code{x1}}{Variable de entrada 1.}
\item{\code{x2}}{Variable de entrada 2.}
\item{\code{y1}}{Variable de salida 1.}
}
}
}
\source{
Los datos fueron creados con propósitos ilustrativos.
}
\usage{
datos3
}
\description{
Este conjunto de datos contiene tres variables: \code{x1}, \code{x2}, \code{y1}.
}
\keyword{datasets}
