% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{GRID}
\alias{GRID}
\title{Construye un objeto de clase GRID.}
\usage{
GRID(data, inputs, outputs, d)
}
\arguments{
\item{data}{Un data.frame que contiene los datos.}

\item{inputs}{Un vector de caracteres que especifica las variables de entrada.}

\item{outputs}{Un vector de caracteres que especifica las variables de salida.}

\item{d}{Un vector numérico que especifica el número de particiones en cada dimensión de las variables de entrada.}
}
\value{
Un objeto de clase GRID.
}
\description{
Esta función crea un objeto de clase GRID que encapsula un conjunto de datos, las variables de entrada y salida,
así como la configuración del grid utilizado para dividir el espacio de entrada.
}
\examples{
data <- data.frame(x1 = c(1, 2, 3), x2 = c(4, 5, 6), y1 = c(7, 8, 9))
inputs <- c("x1", "x2")
outputs <- c("y1")
d <- c(2, 2)
grid <- GRID(data = data, inputs = inputs, outputs = outputs, d = d)
print(grid)
}
