% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{initialize_GRID}
\alias{initialize_GRID}
\title{Función para inicializar la cuadrícula.}
\usage{
initialize_GRID(data, inputs, outputs, d)
}
\arguments{
\item{data}{Los datos para construir la cuadrícula.}

\item{inputs}{Los nombres de las variables de entrada.}

\item{outputs}{Los nombres de las variables de salida.}

\item{d}{La cantidad de dimensiones de la cuadrícula.}
}
\value{
Retorna una lista que representa la cuadrícula inicializada.
}
\description{
Esta función inicializa una cuadrícula con los datos, variables de entrada,
variables de salida y la cantidad de dimensiones especificadas.
}
