% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svf_functions.R
\name{plot_production_frontier}
\alias{plot_production_frontier}
\title{Graficar la frontera de producción}
\usage{
plot_production_frontier(svf, input_col, output_col, d, n_puntos = 100)
}
\arguments{
\item{svf}{Objeto de clase SVF entrenado.}

\item{input_col}{Nombre de la columna del input (x1).}

\item{output_col}{Nombre de la columna del output (y1).}

\item{d}{Valor fijo para los otros inputs que no se están graficando.}

\item{n_puntos}{Número de puntos a utilizar para la estimación.}
}
\description{
Esta función toma un objeto SVF entrenado y grafica la frontera de producción
para un solo input (x1) y un solo output (y1).
}
\examples{
# Definir los datos
data <- data.frame(
  x1 = c(1, 2, 3, 4),
  x2 = c(1, 3, 2, 4),
  y1 = c(1, 3, 2, 4)
)

# Definir listas de inputs, outputs y la cantidad de particiones
inputs <- c("x1", "x2")
outputs <- c("y1")
d <- 3

# Parámetros iniciales del modelo SVF
C <- 1
eps <- 0
method <- 'SSVF'

# Instanciar y entrenar el modelo SVF
svf_instance <- create_SVF(method, inputs, outputs, data, C, eps, d)
svf_instance <- train.SSVF(svf_instance)

# Resolver el modelo y mostrar las soluciones
svf_solution <- solve(svf_instance)

# Imprimir las soluciones de w
print(svf_solution$w)

# Probar la función get_estimation
estimations <- list(
  get_estimation.SVF(svf_instance, c(1, 2)),
  get_estimation.SVF(svf_instance, c(3, 4)),
  get_estimation.SVF(svf_instance, c(1, 7)),
  get_estimation.SVF(svf_instance, c(7, 1)),
  get_estimation.SVF(svf_instance, c(2, 4)),
  get_estimation.SVF(svf_instance, c(7, 1))
)

# Imprimir las estimaciones
for (i in seq_along(estimations)) {
  cat(sprintf("Estimación \%d: \%s\n", i + 1, estimations[[i]]))
}

# Usar la función para graficar la frontera de producción
plot_production_frontier(svf_instance, "x1", "y1", d)
}
