% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svf_functions.R
\name{plot_svf_frontier}
\alias{plot_svf_frontier}
\title{Graficar la frontera del modelo SVF}
\usage{
plot_svf_frontier(svf_instance)
}
\arguments{
\item{svf_instance}{Objeto de la clase SVF entrenado.}
}
\value{
Un gráfico que muestra los datos originales y la frontera del modelo SVF.
}
\description{
Esta función genera un gráfico que muestra los datos originales y la
frontera generada por el modelo SVF.
}
