% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{search_dmu}
\alias{search_dmu}
\title{Función para buscar una observación en el grid}
\usage{
search_dmu(grid, dmu)
}
\arguments{
\item{grid}{Un objeto de clase GRID que define la estructura del grid.}

\item{dmu}{Un vector que representa la observación a buscar en el grid.}
}
\value{
Un vector con la posición de la observación en el grid.
}
\description{
Esta función devuelve la celda en la que se encuentra una observación en el grid.
}
\examples{
data <- data.frame(x1 = c(1, 2, 3), x2 = c(4, 5, 6), y1 = c(7, 8, 9))
inputs <- c("x1", "x2")
outputs <- c("y1")
d <- c(2, 2)
grid <- GRID(data = data, inputs = inputs, outputs = outputs, d = d)

# Buscar una observación en el grid
search_dmu(grid, c(2, 5))
}
