% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvf.R
\name{train.SSVF}
\alias{train.SSVF}
\title{Entrena el modelo SSVF}
\usage{
train.SSVF(svf)
}
\arguments{
\item{svf}{Objeto SSVF.}
}
\value{
El objeto SSVF con el modelo resuelto añadido.
}
\description{
Esta función configura y resuelve el modelo SSVF utilizando programación lineal mixta.
}
\examples{
source("~/Documents/GitHub/svfpackage/R/ssvf.R")

# Usar datos de prueba
data <- read.table("~/Documents/GitHub/svfpackage/data/datos.txt", header = TRUE, sep = ";")

# Definición de inputs, outputs y otros parámetros
inputs <- c("x1", "x2")
outputs <- c("y1")
d <- 2
C <- 1
eps <- 0
method <- 'SSVF'

# Crear y mostrar el objeto SVF
ssvf <- SSVF(method, inputs, outputs, data, C, eps, d)

trained_svf <- train.SSVF(ssvf)

print(trained_svf)
}
