% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{transformation}
\alias{transformation}
\title{Función para realizar una transformación}
\usage{
transformation(x_i, t_k)
}
\arguments{
\item{x_i}{Valor de la observación a evaluar.}

\item{t_k}{Valor del nodo con el que se quiere comparar.}
}
\value{
-1 si x_i < t_k, 0 si x_i == t_k, 1 si x_i > t_k.
}
\description{
Esta función evalúa si el valor de una observación es mayor, igual o menor que el valor de un nodo del grid.
}
\examples{
# Evaluar una observación menor que el valor del nodo del grid
transformation(2, 5)
# Output: -1

# Evaluar una observación igual al valor del nodo del grid
transformation(3, 3)
# Output: 0

# Evaluar una observación mayor que el valor del nodo del grid
transformation(7, 4)
# Output: 1
}
