% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{GRID}
\alias{GRID}
\title{Clase GRID para modelado SVF}
\usage{
GRID(data, inputs, outputs, d)
}
\description{
Esta clase representa un grid sobre el que se realiza el modelo SVF.
Un grid es una partición del espacio de los inputs dividido por celdas.
}
\section{Fields}{

\describe{
\item{\code{data}}{DataFrame con el conjunto de datos sobre los que se construye el grid.}

\item{\code{inputs}}{Vector de nombres de los inputs.}

\item{\code{outputs}}{Vector de nombres de los outputs.}

\item{\code{d}}{Vector con el número de particiones en las que se divide el grid.}

\item{\code{data_grid}}{Grid de datos, inicialmente NULL.}

\item{\code{knot_list}}{Lista de nodos del grid, inicialmente NULL.}
}}

\examples{
source("~/svfpackage/R/grid.R")

# Crear el dataset de ejemplo
data <- data.frame(x1 = c(1, 2, 3), x2 = c(4, 5, 6), y1 = c(7, 8, 9))

# Definir listas de inputs, outputs y la cantidad de particiones
inputs <- c("x1", "x2")
outputs <- c("y1")
d <- 2

# Inicializar la cuadrícula
grid_instance <- initialize_GRID(data, inputs, outputs, d)
print(grid_instance)
}
