% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{search_dmu}
\alias{search_dmu}
\title{Función para buscar el índice de la celda en la cuadrícula asociada a un punto dado.}
\usage{
search_dmu(grid_instance, dmu)
}
\arguments{
\item{grid_instance}{La instancia de la cuadrícula.}

\item{dmu}{La lista de valores que representan un punto en la cuadrícula.}
}
\value{
Retorna una lista con los índices de celda correspondientes a las dimensiones de la cuadrícula.
}
\description{
Esta función busca el índice de la celda en la cuadrícula asociada a un punto
dado utilizando el método de transformación definido en la función \code{transformation}.
}
