% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svfgrid.R
\name{SVFGrid}
\alias{SVFGrid}
\title{Constructor para la clase SVFGrid}
\usage{
SVFGrid(data, inputs, outputs, d)
}
\arguments{
\item{data}{DataFrame con los datos sobre los cuales se construirá el grid.}

\item{inputs}{Vector con los nombres de las columnas que serán tratadas como inputs.}

\item{outputs}{Vector con los nombres de las columnas que serán tratadas como outputs.}

\item{d}{Entero o vector de enteros que especifica el número de divisiones por dimensión en el grid.}
}
\value{
Un objeto de clase SVFGrid.
}
\description{
Esta función crea una instancia de la clase SVFGrid, la cual es una
extensión de la clase GRID con funcionalidades adicionales específicas
para el manejo de grids en el contexto de análisis SVF.
}
\examples{
source("~/Documents/GitHub/svfpackage/R/svfgrid.R")

# Crear un conjunto de datos de ejemplo
data <- data.frame(x1 = c(1, 2, 3, 4), x2 = c(5, 6, 7, 8), y1 = c(9, 1, 2, 3))

# Definir listas de inputs, outputs y la cantidad de particiones
inputs <- c("x1", "x2")
outputs <- c("y1")
d <- 2

# Crear la instancia de la clase SVFGrid y llamar al método create_grid
grid_obj <- SVFGrid(data, inputs, outputs, d)
print(grid_obj)
}
