% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{transformation}
\alias{transformation}
\title{Transforma un valor basado en un nodo en la cuadrícula.}
\usage{
transformation(x_i, t_k)
}
\arguments{
\item{x_i}{El valor que se va a transformar.}

\item{t_k}{El nodo en la cuadrícula.}
}
\value{
Retorna -1 si x_i es menor que t_k, 0 si son iguales, y 1 si x_i es mayor que t_k.
}
\description{
Esta función toma un valor y lo compara con un nodo en la cuadrícula.
Dependiendo de la relación entre el valor y el nodo, se devuelve -1, 0 o 1.
}
\examples{
transformation(2, 5)
transformation(3, 3)
transformation(7, 4)
}
